src/pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Võ Đình Đạt <vodinhdat03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gradio = "^4.21.0"
einops = "^0.7.0"
python-dotenv = "^1.0.1"
sentence-transformers = "^2.5.1"
pydantic = "^2.6.4"
llama-index = "^0.10.20"
llama-index-embeddings-huggingface = "^0.1.4"
llama-index-retrievers-bm25 = "^0.1.3"
llama-index-llms-openai = "^0.1.12"
llama-index-vector-stores-chroma = "^0.1.6"
chromadb = "^0.4.24"
llama-index-llms-ollama = "^0.1.2"
llama-index-readers-file = "^0.1.11"
pymupdf = "^1.23.26"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
